r/aws_iam_saml_provider: Use 'samlProviderCreateTags'.
