+++ internal/service/iam/saml_provider.go
@@ -12,9 +12,11 @@ import (
 	"github.com/aws/aws-sdk-go/service/iam"
 	"github.com/hashicorp/aws-sdk-go-base/v2/awsv1shim/v2/tfawserr"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
+	"github.com/hashicorp/terraform-provider-aws/internal/errs"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
 	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
 	"github.com/hashicorp/terraform-provider-aws/internal/verify"
@@ -66,18 +68,16 @@ func resourceSAMLProviderCreate(ctx context.Context, d *schema.ResourceData, met
 	conn := meta.(*conns.AWSClient).IAMConn()
 
 	name := d.Get("name").(string)
-	tags := GetTagsIn(ctx)
 	input := &iam.CreateSAMLProviderInput{
 		Name:                 aws.String(name),
 		SAMLMetadataDocument: aws.String(d.Get("saml_metadata_document").(string)),
-		Tags:                 tags,
+		Tags:                 GetTagsIn(ctx),
 	}
 
 	output, err := conn.CreateSAMLProviderWithContext(ctx, input)
 
-	// Some partitions (i.e., ISO) may not support tag-on-create
-	if input.Tags != nil && verify.ErrorISOUnsupported(conn.PartitionID, err) {
-		log.Printf("[WARN] failed creating IAM SAML Provider (%s) with tags: %s. Trying create without tags.", name, err)
+	// Some partitions (e.g. ISO) may not support tag-on-create.
+	if input.Tags != nil && errs.IsUnsupportedOperationInPartitionError(conn.PartitionID, err) {
 		input.Tags = nil
 
 		output, err = conn.CreateSAMLProviderWithContext(ctx, input)
@@ -89,18 +89,17 @@ func resourceSAMLProviderCreate(ctx context.Context, d *schema.ResourceData, met
 
 	d.SetId(aws.StringValue(output.SAMLProviderArn))
 
-	// Some partitions (i.e., ISO) may not support tag-on-create, attempt tag after create
-	if input.Tags == nil && len(tags) > 0 {
-		err := samlProviderUpdateTags(ctx, conn, d.Id(), nil, KeyValueTags(ctx, tags))
+	// For partitions not supporting tag-on-create, attempt tag after create.
+	if tags := GetTagsIn(ctx); input.Tags == nil && len(tags) > 0 {
+		err := samlProviderCreateTags(ctx, conn, d.Id(), tags)
 
-		// If default tags only, log and continue. Otherwise, error.
-		if v, ok := d.GetOk("tags"); (!ok || len(v.(map[string]interface{})) == 0) && verify.ErrorISOUnsupported(conn.PartitionID, err) {
-			log.Printf("[WARN] failed adding tags after create for IAM SAML Provider (%s): %s", d.Id(), err)
+		// If default tags only, continue. Otherwise, error.
+		if v, ok := d.GetOk(names.AttrTags); (!ok || len(v.(map[string]interface{})) == 0) && errs.IsUnsupportedOperationInPartitionError(conn.PartitionID, err) {
 			return resourceSAMLProviderRead(ctx, d, meta)
 		}
 
 		if err != nil {
-			return diag.Errorf("adding tags after create for IAM SAML Provider (%s): %s", d.Id(), err)
+			return diag.Errorf("setting IAM SAML Provider (%s) tags: %s", d.Id(), err)
 		}
 	}
 
@@ -151,7 +150,6 @@ func resourceSAMLProviderUpdate(ctx context.Context, d *schema.ResourceData, met
 			SAMLMetadataDocument: aws.String(d.Get("saml_metadata_document").(string)),
 		}
 
-		log.Printf("[DEBUG] Updating IAM SAML Provider: %s", input)
 		_, err := conn.UpdateSAMLProviderWithContext(ctx, input)
 
 		if err != nil {
@@ -164,9 +162,8 @@ func resourceSAMLProviderUpdate(ctx context.Context, d *schema.ResourceData, met
 
 		err := samlProviderUpdateTags(ctx, conn, d.Id(), o, n)
 
-		// Some partitions (i.e., ISO) may not support tagging, giving error
-		if verify.ErrorISOUnsupported(conn.PartitionID, err) {
-			log.Printf("[WARN] failed updating tags for IAM SAML Provider (%s): %s", d.Id(), err)
+		// Some partitions (e.g. ISO) may not support tagging.
+		if errs.IsUnsupportedOperationInPartitionError(conn.PartitionID, err) {
 			return resourceSAMLProviderRead(ctx, d, meta)
 		}
 
@@ -195,6 +192,31 @@ func resourceSAMLProviderDelete(ctx context.Context, d *schema.ResourceData, met
 	}
 
 	return nil
+}
+
+func FindSAMLProviderByARN(ctx context.Context, conn *iam.IAM, arn string) (*iam.GetSAMLProviderOutput, error) {
+	input := &iam.GetSAMLProviderInput{
+		SAMLProviderArn: aws.String(arn),
+	}
+
+	output, err := conn.GetSAMLProviderWithContext(ctx, input)
+
+	if tfawserr.ErrCodeEquals(err, iam.ErrCodeNoSuchEntityException) {
+		return nil, &retry.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	return output, nil
 }
 
 func nameFromSAMLProviderARN(v string) (string, error) {
